backend/pyproject.toml
[project]
name = "backend-jwt"
version = "0.1.0"
description = "FastAPI JWT authentication service"
authors = [
    {name = "TomiGiacobino"}
]
requires-python = ">=3.11"
dependencies = [
    "fastapi (>=0.136.1,<0.137.0)",
    "uvicorn[standard] (>=0.46.0,<0.47.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-multipart (>=0.0.28,<0.0.29)",
    "bcrypt (>=3.0.0,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
